changed float dtype to 32 bit floats

+++ particle_swarm/algorithm/particle_swarm.py
@@ -1,5 +1,6 @@
 """Program entry point."""
 
+import numba
 import numpy as np
 
 from particle_swarm.algorithm import runner
@@ -64,7 +65,7 @@ def build_swarm(min_pos, max_pos):
 
 def _build_particle(min_pos, max_pos):
 
-    pos = random_generator.uniform(low=min_pos, high=max_pos, size=DIMENSIONS)
+    pos = random_generator.uniform(low=min_pos, high=max_pos, size=DIMENSIONS).astype(np.float32)
     vel = pos
 
     return {
